+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - translation fixes [amleczko]
 - portlet hidden for anonymous [amleczko]
+- make sure only published content is used [amleczko]
 
 
 1.0 (2013-03-05)